backend/requirements.txt
# Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0

# Cache
redis==5.0.1

# Utils
python-dotenv==1.0.1
httpx==0.26.0

# Dev
pytest==8.0.0
pytest-asyncio==0.23.4
aiosqlite==0.19.0
